+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project that require documentation updates will be d
 
 ## [NEXT RELEASE]
 - Added a global flag `--token-file` to roxctl causing an API token to be read from the specified file (ROX-2319).
+- Emit warnings if defaults for `--create-updater-sa` or `--collection-method` are used for roxctl (ROX-5264).
 
 ## [46.0]
 - Added the following REST APIs: